requirements.txt
# FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0
python-dotenv>=1.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
sqlite3-fts4

# Award Flight APIs & Data
httpx>=0.25.0
airportsdata>=20240719
python-dateutil>=2.8.2

# Core dependencies
selenium>=4.9.0
undetected-chromedriver>=3.5.0
webdriver-manager>=4.0.0
scrapy>=2.11.0

# Utilities
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
pandas>=2.0.0
fake-useragent>=1.1.3
psutil>=5.9.0

# Data processing
pytz>=2023.3

# Networking
urllib3>=2.0.0
aiohttp>=3.8.5

# Logging
loguru>=0.7.0 